Procedural Generation Code
Run this as a project in unity to get a randomly generated map program. Also contains the .exe file. use wasd or arrow keys to look around and scroll wheel to zoom in and out. The code isnt yet fully documented so if you dont understand anything just ask. Or if you do understand something add comments explaining it I guess.