Changing Sprites & 'Building' Buildings
Added a new SpriteRenderer to the Tile class so that selecting and hovering over tiles is now indipendent from the sprite rendered based on the type of tile.
Added a new House sprite that is placed when a selected tile is clicked again.
Removed obselete tile sprites (the 'selected' and 'hover' variants).